+++ prototype/src/main/java/com/wantensoup/prototype/Menu/MenuRepository.java
@@ -9,7 +9,7 @@ import org.springframework.stereotype.Repository;
  */
 
 @Repository
-public interface MenuRepository extends JpaRepository<Menu, Long>{
+public interface MenuRepository extends JpaRepository<Menu, Integer>{
     
     public Menu findByMenuItem(String menuItem);
