+++ prototype/src/main/java/com/wantensoup/prototype/Schedule/ScheduleRepository.java
@@ -10,7 +10,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ScheduleRepository extends JpaRepository<Schedule, Long> {
+public interface ScheduleRepository extends JpaRepository<Schedule, Integer> {
 
     public Schedule findByEmployeeName(String _employeeName);
